Update to cleaner router.
